Create and load material files
- Add 64bit version to asset file header
- Convert material asset coordinate properties to int8_t
- Add path in asset data
- Load material assets from file
- Create material files from GLTF materials
- Load materials from engine specific files

+++ engine/src/liquid/asset/MaterialAsset.h
@@ -4,24 +4,24 @@ namespace liquid {
 
 struct MaterialAsset {
   TextureAssetHandle baseColorTexture = TextureAssetHandle::Invalid;
-  int baseColorTextureCoord = -1;
+  int8_t baseColorTextureCoord = -1;
   glm::vec4 baseColorFactor{};
 
   TextureAssetHandle metallicRoughnessTexture = TextureAssetHandle::Invalid;
-  int metallicRoughnessTextureCoord = -1;
+  int8_t metallicRoughnessTextureCoord = -1;
   float metallicFactor = 0.0f;
   float roughnessFactor = 0.0f;
 
   TextureAssetHandle normalTexture = TextureAssetHandle::Invalid;
-  int normalTextureCoord = -1;
+  int8_t normalTextureCoord = -1;
   float normalScale = 0.0f;
 
   TextureAssetHandle occlusionTexture = TextureAssetHandle::Invalid;
-  int occlusionTextureCoord = -1;
+  int8_t occlusionTextureCoord = -1;
   float occlusionStrength = 0.0f;
 
   TextureAssetHandle emissiveTexture = TextureAssetHandle::Invalid;
-  int emissiveTextureCoord = -1;
+  int8_t emissiveTextureCoord = -1;
   glm::vec3 emissiveFactor{};
 };
